Compliance Policies
Problem:
Interface to compliance endpoint is unavailable

Solution:
Added Compliance Policies endpoint and logic to detemine
compliance for one asset. A compliance_demo python script
is available for testing.

+++ scripts/unittests.sh
@@ -4,6 +4,23 @@
 #
 python3 --version
 
+# run single test from cmdline for example:
+#
+# all tests in a module (name of file in unittest directory):
+#       UNITTEST=testcompliance_policy_type task unittests
+#
+# all tests in a class
+#       UNITTEST=testcompliance_policy_type.TestCompliancePolicyType task unittests
+#
+# single test
+#       UNITTEST=testcompliance_policy_type.TestCompliancePolicyType.test_compliance_policy_type task unittests
+#
+if [ -n "${UNITTEST}" ]
+then
+    python3 -m unittest -v unittests.${UNITTEST}
+    exit 0
+fi
+
 rm -f coverage.xml
 rm -rf htmlcov
 COVERAGE="coverage"
